Well, the module cleaner section is a bit of a disaster.
It seems to work fine for most of the modules, even when dumping all the modules and having it re-insert them all, but there are a few places where it screws the module order up (mostly an issue because of the main pages shown on ACP pages when visiting a main category), in a few places it misses adding items and doesn't recognize that they are missing (can not figure this one out), and it gives an error about some modules already existing when adding them all.

I think I'll leave this for now, but if nobody can figure out a better way to do it or how to fix it I'd suggest either dropping the modules part or only having the option to completely reset the modules and not the add missing/remove extra I have been trying to do.