SQL Query - This new way looks much nicer for results.
Changed the section for getting a default in the build_cfg_template.  Give the option to force it to not get the default through request_var and never get a default for password fields (unless explicitly set in $vars['default']).

tool_options.html - Stop Firefox from trying to fill everything in automatically.