Fixing an issue as described by Patrick:
The SRT Generator doesn't report a version number. This is because it's marked as a dropdown and the prefilled value gets erased from the answer array when it tries to load the value from the options array. It actually throws an error about an undefined index, but you can't see it unless you step through it in a debugger.
"[phpBB Debug] PHP Notice: in file /stk/tools/main/srt_generator.php on line 348: Undefined index: 3.0.7-PL1"

It looks like you were going to present them with the option of choosing between the two if they were different.

#62280