[issue/BOM] Remove the BOM character
Your config.php has been saved in the "UTF-8 WITH BOM" character
set, this adds the "BOM" character to the beginning of the file
which potentially breaks phpBB/the STK, therefore this file
triggers the BOM sniffer.
All phpBB/STK files should be saved as "UTF-8 WITHOUT BOM",
I've corrected the file for you.