[feature/support-phpBB-3.0.9] Potentially corrupted database in firebird
Add the Q&A captcha tables to the schema if at least one of the tables is found, because in that case we assume that the user had installed that captcha.
A special case if when the user is using the firebird DBMS *and* he is running phpBB < 3.0.9, because due to a bug in phpBB the Q&A captcha couldn't be correctly installed in that case thus don't include these tables in the schema so that the cleaner can delete them if needed.
The worker logic will adjust the config data if needed.
More info see: https://github.com/phpbb/phpbb3/blob/release-3.0.9-RC2/phpBB/install/database_update.php#L1928