[prep-release/1.0.5] Fix STK login
The `$phpbb_root_path` gets incorrectly rewritten to resolve a potential
invalid link (pointing to phpBB rather than the STK index). As this only
occures on error messages and all other redirects remain in working order
remove this overwritten behavior for now. This issue should be addressed
for the next release but poses no major problems for now.